ref: move credentials into the backend
