--- autogpt_platform/autogpt_libs/autogpt_libs/integration_credentials_store/__init__.py
@@ -1,9 +1,0 @@
-from .store import IntegrationCredentialsStore
-from .types import Credentials, APIKeyCredentials, OAuth2Credentials
-
-__all__ = [
-    "IntegrationCredentialsStore",
-    "Credentials",
-    "APIKeyCredentials",
-    "OAuth2Credentials",
-]
