+++ autogpt_platform/backend/backend/blocks/ai_music_generator.py
@@ -4,11 +4,15 @@ from enum import Enum
 from typing import Literal
 
 import replicate
-from autogpt_libs.integration_credentials_store.types import APIKeyCredentials
 from pydantic import SecretStr
 
 from backend.data.block import Block, BlockCategory, BlockOutput, BlockSchema
-from backend.data.model import CredentialsField, CredentialsMetaInput, SchemaField
+from backend.data.model import (
+    APIKeyCredentials,
+    CredentialsField,
+    CredentialsMetaInput,
+    SchemaField,
+)
 
 logger = logging.getLogger(__name__)
