+++ autogpt_platform/backend/backend/blocks/discord.py
@@ -3,11 +3,15 @@ from typing import Literal
 
 import aiohttp
 import discord
-from autogpt_libs.integration_credentials_store.types import APIKeyCredentials
 from pydantic import SecretStr
 
 from backend.data.block import Block, BlockCategory, BlockOutput, BlockSchema
-from backend.data.model import CredentialsField, CredentialsMetaInput, SchemaField
+from backend.data.model import (
+    APIKeyCredentials,
+    CredentialsField,
+    CredentialsMetaInput,
+    SchemaField,
+)
 
 DiscordCredentials = CredentialsMetaInput[Literal["discord"], Literal["api_key"]]
