+++ autogpt_platform/backend/backend/blocks/github/_auth.py
@@ -1,12 +1,13 @@
 from typing import Literal
 
-from autogpt_libs.integration_credentials_store.types import (
-    APIKeyCredentials,
-    OAuth2Credentials,
-)
 from pydantic import SecretStr
 
-from backend.data.model import CredentialsField, CredentialsMetaInput
+from backend.data.model import (
+    APIKeyCredentials,
+    CredentialsField,
+    CredentialsMetaInput,
+    OAuth2Credentials,
+)
 from backend.util.settings import Secrets
 
 secrets = Secrets()
