+++ autogpt_platform/backend/backend/blocks/google/_auth.py
@@ -1,9 +1,8 @@
 from typing import Literal
 
-from autogpt_libs.integration_credentials_store.types import OAuth2Credentials
 from pydantic import SecretStr
 
-from backend.data.model import CredentialsField, CredentialsMetaInput
+from backend.data.model import CredentialsField, CredentialsMetaInput, OAuth2Credentials
 from backend.util.settings import Secrets
 
 # --8<-- [start:GoogleOAuthIsConfigured]
