+++ autogpt_platform/backend/backend/blocks/ideogram.py
@@ -1,12 +1,16 @@
 from enum import Enum
 from typing import Any, Dict, Literal, Optional
 
-from autogpt_libs.integration_credentials_store.types import APIKeyCredentials
 from pydantic import SecretStr
 from requests.exceptions import RequestException
 
 from backend.data.block import Block, BlockCategory, BlockOutput, BlockSchema
-from backend.data.model import CredentialsField, CredentialsMetaInput, SchemaField
+from backend.data.model import (
+    APIKeyCredentials,
+    CredentialsField,
+    CredentialsMetaInput,
+    SchemaField,
+)
 from backend.util.request import requests
 
 TEST_CREDENTIALS = APIKeyCredentials(
