+++ autogpt_platform/backend/backend/blocks/replicate_flux_advanced.py
@@ -3,12 +3,16 @@ from enum import Enum
 from typing import Literal
 
 import replicate
-from autogpt_libs.integration_credentials_store.types import APIKeyCredentials
 from pydantic import SecretStr
 from replicate.helpers import FileOutput
 
 from backend.data.block import Block, BlockCategory, BlockOutput, BlockSchema
-from backend.data.model import CredentialsField, CredentialsMetaInput, SchemaField
+from backend.data.model import (
+    APIKeyCredentials,
+    CredentialsField,
+    CredentialsMetaInput,
+    SchemaField,
+)
 
 TEST_CREDENTIALS = APIKeyCredentials(
     id="01234567-89ab-cdef-0123-456789abcdef",
