+++ autogpt_platform/backend/backend/blocks/talking_head.py
@@ -1,11 +1,15 @@
 import time
 from typing import Literal
 
-from autogpt_libs.integration_credentials_store.types import APIKeyCredentials
 from pydantic import SecretStr
 
 from backend.data.block import Block, BlockCategory, BlockOutput, BlockSchema
-from backend.data.model import CredentialsField, CredentialsMetaInput, SchemaField
+from backend.data.model import (
+    APIKeyCredentials,
+    CredentialsField,
+    CredentialsMetaInput,
+    SchemaField,
+)
 from backend.util.request import requests
 
 TEST_CREDENTIALS = APIKeyCredentials(
