+++ autogpt_platform/backend/backend/blocks/text_to_speech_block.py
@@ -1,10 +1,14 @@
 from typing import Any, Literal
 
-from autogpt_libs.integration_credentials_store.types import APIKeyCredentials
 from pydantic import SecretStr
 
 from backend.data.block import Block, BlockCategory, BlockOutput, BlockSchema
-from backend.data.model import CredentialsField, CredentialsMetaInput, SchemaField
+from backend.data.model import (
+    APIKeyCredentials,
+    CredentialsField,
+    CredentialsMetaInput,
+    SchemaField,
+)
 from backend.util.request import requests
 
 TEST_CREDENTIALS = APIKeyCredentials(
