+++ autogpt_platform/backend/backend/data/__init__.py
@@ -1,0 +1,9 @@
+from .credential_store import IntegrationCredentialsStore
+from .model import APIKeyCredentials, Credentials, OAuth2Credentials
+
+__all__ = [
+    "IntegrationCredentialsStore",
+    "Credentials",
+    "APIKeyCredentials",
+    "OAuth2Credentials",
+]
