+++ autogpt_platform/backend/backend/integrations/creds_manager.py
@@ -2,14 +2,12 @@ import logging
 from contextlib import contextmanager
 from datetime import datetime
 
-from autogpt_libs.integration_credentials_store import (
-    Credentials,
-    IntegrationCredentialsStore,
-)
 from autogpt_libs.utils.synchronize import RedisKeyedMutex
 from redis.lock import Lock as RedisLock
 
 from backend.data import redis
+from backend.data.credential_store import IntegrationCredentialsStore
+from backend.data.model import Credentials
 from backend.integrations.oauth import HANDLERS_BY_NAME, BaseOAuthHandler
 from backend.util.settings import Settings
