+++ autogpt_platform/backend/backend/integrations/oauth/github.py
@@ -2,8 +2,7 @@ import time
 from typing import Optional
 from urllib.parse import urlencode
 
-from autogpt_libs.integration_credentials_store import OAuth2Credentials
-
+from backend.data.model import OAuth2Credentials
 from backend.util.request import requests
 
 from .base import BaseOAuthHandler
